session voters refactoring.
