+++ src/Ilios/AuthenticationBundle/Resources/config/dto_voters.yml
@@ -5,3 +5,9 @@ services:
         public: false
         tags:
             - { name: security.voter }
+    security.access.dto.sesision_voter:
+        class: Ilios\AuthenticationBundle\Voter\DTO\SessionDTOVoter
+        parent: security.access.course_voter
+        public: false
+        tags:
+            - { name: security.voter }
