+++ src/Ilios/AuthenticationBundle/Resources/config/entity_voters.yml
@@ -5,3 +5,9 @@ services:
         public: false
         tags:
             - { name: security.voter }
+    security.access.entity.session_voter:
+        class: Ilios\AuthenticationBundle\Voter\Entity\SessionEntityVoter
+        parent: security.access.course_voter
+        public: false
+        tags:
+            - { name: security.voter }
